requirements.txt
bcrypt==3.1.4
bottle==0.12.13
cffi==1.11.5
itsdangerous==0.24
peewee==3.5.0
pycparser==2.18
PyJWT==1.6.4
six==1.11.0
